chore: delete formik validation
