+++ .eslintrc.json
@@ -12,7 +12,8 @@
     "@typescript-eslint/ban-types": "off",
     "@typescript-eslint/no-unused-vars": "off",
 //    "react/react-in-jsx-scope": "off"
-    "react/display-name": "off"
+    "react/display-name": "off",
+    "no-useless-escape": "warn"
   },
   "settings": {
     "react": {
